% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{all_true}
\alias{all_true}
\title{Are all elements TRUE?}
\usage{
all_true(x)
}
\arguments{
\item{x}{A vector, matrix or array.}
}
\value{
Boolean scalar. TRUE if every element of \code{x} is TRUE.
}
\description{
Check whether all elements of a vector, matrix or array are TRUE.
}
\details{
Note that numeric values are not coerced to logical. See examples.
}
\examples{
x <- array(TRUE, c(1, 2, 3))
all_true(x)

x[1, 2, 3] <- FALSE
all_true(x)

# Numeric is not coerced to logical
x[1, 2, 3] <- 1
all_true(x)

# A list might fail
y <- list(a = TRUE, b = matrix(TRUE, 3, 3), c = list(TRUE, TRUE))
all_true(y)
}
