% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{big_A}
\alias{big_A}
\alias{big_Y}
\alias{big_nu}
\alias{big_U}
\alias{big_SIGMA}
\title{Convert objects to VAR(1) companion format}
\usage{
big_A(A)

big_Y(Y, p)

big_nu(nu, p)

big_U(U, p)

big_SIGMA(SIGMA, p)
}
\arguments{
\item{A}{A \code{(K x Kp)} matrix, providing the coeffcients for lag \code{1} to \code{p}
with the first row containing the coefficents of the first equation.
Parameter \code{p} is the maximum lag length and \code{K} the number of variables.}

\item{p}{An integer scalar. The lag length of the VAR(p) system.}

\item{nu}{A \code{(K x 1)} matrix of intercepts.}

\item{U}{A \code{(K x N)} matrix of residuals.}

\item{SIGMA}{A \code{(K x K)} matrix of covariances. The covariance matrix of the
residuals \code{U}.}
}
\value{
\itemize{
\item \code{big_A} \cr A \code{(Kp x Kp)} matrix with the slope parameters on the
first K rows. The remaining rows carry block-diagonal unit matrices and the
remainder are zeros.
}

\itemize{
\item \code{big_Y} \cr A \code{(Kp x N-p+1)} matrix. The \code{p-1} lags of \code{Y} are
pasted as rows below \code{Y}. This leads to the loss of \code{p-1} sample
observations.
}

\itemize{
\item \code{big_nu} \cr A \code{(Kp x 1)} colum matrix. The first \code{K} elements
correspond to \code{nu}. The remaining elements are zero.
}

\itemize{
\item \code{big_U} \cr A \code{(Kp x N)} matrix. The first \code{K} rows contain the
residuals, the last \code{K(p-1)} rows consist of zeros. TODO: does not really
correspond to the output of \code{big_Y}. There, it's \code{N-p+1} columns; here,
it's \code{N}.
}

\itemize{
\item \code{big_SIGMA} \cr A \code{(Kp x Kp)} matrix. The upper left \code{(K x K)}
block will contain the original covariance matrix: The remaining elements
will be zero.
}
}
\description{
Convert objects such as a matrix of slope parameters \code{A} or intercepts \code{nu},
a matrix of observations \code{Y} or residuals \code{U} or a covariance matrix \code{SIGMA}
such that they correspond to objects taken from a VAR(1) representation.
}
\section{Note}{
 difference between big_Y and Y2Z? matrix \code{Z} has \code{K*p + 1}
rows whereas the output of \code{big_Y} has \code{K*p} rows. The very last lag is
missing.
}

\examples{
K <- 4
N <- 7
p <- 2

A <- matrix(0.1, K, K * p)
big_A(A)

Y <- matrix(seq_len(K * N), K, N)
big_Y(Y, p)

nu <- as.matrix(seq_len(K))
big_nu(nu, p)

U <- matrix(seq_len(K * N), K, N)
big_U(U, p)

SIGMA <- matrix(0.5, K, K)
big_SIGMA(SIGMA, p = p)
}
