% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{check_stability}
\alias{check_stability}
\title{Check the stability criterion of a VAR(p) model}
\usage{
check_stability(A)
}
\arguments{
\item{A}{A \code{(K x Kp)} matrix, providing the coeffcients for lag \code{1} to \code{p}
with the first row containing the coefficents of the first equation.
Parameter \code{p} is the maximum lag length and \code{K} the number of variables.}
}
\value{
Boolean scalar. \code{TRUE} if the criterion is satisfied, otherwise
\code{FALSE}.
}
\description{
This function will check whether the process implied by the coefficient
matrix \code{A} is stable. That is, whether the eigenvalues of the associated
companion format are all smaller than one in absolute terms.
}
\examples{
A <- matrix(c(.5, .4, .1, .5, 0, .25, 0, 0), nrow = 2)
check_stability(A)

A <- matrix(c(1, 0, 0, 1), nrow = 2)
check_stability(A)

A <- matrix(c(1, 0, 0, 1, -0.1, 0, 0, -0.1), nrow = 2)
check_stability(A)
}
