% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-2-estimation.R
\name{chol_decomp}
\alias{chol_decomp}
\title{Cholesky Decomposition}
\usage{
chol_decomp(A)
}
\arguments{
\item{A}{A matrix, must be positive definite. If not, an error will be
thrown.}
}
\value{
A matrix, the 'square root' of matrix \code{A}, as in \eqn{A =
  LL^{T}}{A = LL'}. TODO IS IT L or L'?
}
\description{
Cholesky decomposition of a positive-definite matrix using the
Cholesky-Banachiewicz algorithm.
}
