% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{commutation_matrix}
\alias{commutation_matrix}
\title{Title}
\usage{
commutation_matrix(r, c)
}
\arguments{
\item{c}{}
}
\description{
Title
}
\examples{
r <- 3
c <- 2

A <- matrix(rnorm(r * c), r, c)

K <- commutation_matrix(r, c)
all(K \%*\% vec(A) == vec(t(A)))
}
