% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{companion_format}
\alias{companion_format}
\alias{big_A}
\alias{big_Y}
\alias{big_nu}
\alias{big_U}
\alias{big_SIGMA}
\title{Convert a VAR(p) model to VAR(1) companion format}
\usage{
companion_format(Y, nu, A, U)

big_A(A)

big_Y(Y, p)

big_nu(nu, p)

big_U(U, p)

big_SIGMA(SIGMA, p)
}
\arguments{
\item{Y}{A \code{(K x N+p)} matrix carrying the data for estimation. There are
\code{N} observations for each of the \code{K} variables with \code{p} pre-sample values.}

\item{nu}{A \code{(K x 1)} matrix of intercepts.}

\item{A}{A \code{(K x Kp)} matrix, providing the coeffcients for lag \code{1} to \code{p}
with the first row containing the coefficents of the first equation.
Parameter \code{p} is the maximum lag length and \code{K} the number of variables.}

\item{U}{A \code{(K x N)} matrix of residuals.}

\item{p}{An integer scalar. The lag length of the VAR(p) system.}

\item{SIGMA}{A \code{(K x K)} matrix of covariances. The covariance matrix of the
residuals \code{U}.}
}
\value{
\itemize{
\item \code{companion_format} \cr A list with five elements. Four elements
\code{Y}, \code{nu}, \code{A} and \code{U} correspond to the output of \code{big_Y}, \code{big_nu},
\code{big_A}, and \code{big_U}. The fifth element is the \code{(Kp x N)} regressor matrix
\code{Z} populated with lags of \code{Y}.
}

\itemize{
\item \code{big_A} \cr A \code{(Kp x Kp)} matrix with the slope parameters on the
first K rows. The remaining rows carry block-diagonal unit matrices and the
remainder are zeros.
}

\itemize{
\item \code{big_Y} \cr A \code{(Kp x N)} matrix. The \code{p-1} lags of \code{Y} are
pasted as rows below \code{Y}. The \code{p} pre-sample observations are deleted.
}

\itemize{
\item \code{big_nu} \cr A \code{(Kp x 1)} colum matrix. The first \code{K} elements
correspond to \code{nu}. The remaining elements are zero.
}

\itemize{
\item \code{big_U} \cr A \code{(Kp x N)} matrix. The first \code{K} rows contain the
residuals, the last \code{K(p-1)} rows consist of zeros.
}

\itemize{
\item \code{big_SIGMA} \cr A \code{(Kp x Kp)} matrix. The upper left \code{(K x K)}
block will contain the original covariance matrix. The remaining elements
will be zero.
}
}
\description{
Convert matrices containig observations, coefficients, residuals and
covariances to their VAR(1) companion format.
}
\details{
Convert objects such as
\itemize{
\item \code{A} - a matrix of slope parameters
\item \code{nu} -  a column vector of intercepts
\item \code{Y} - a matrix of observations
\item \code{U} - a matrix of  residuals
\item \code{SIGMA} - a covariance matrix
}

such that they correspond to objects taken from a
VAR(1) representation.
}
\section{Note}{
 The difference between big_Y() and \code{\link[=Y2Z]{Y2Z()}} is that
observations in matrix \code{Z} are shifted back by one time period in
comparison to \code{Y}.
}

\examples{
set.seed(8191)

K <- 3
N <- 5
p <- 2

nu <- matrix(1:K, ncol = 1)
A  <- matrix(0.1, K, K * p); diag(A) <- 1:K / 10
U  <- matrix(rnorm(K * N), K, N)
Y0 <- matrix(0, K, p)
Y  <- create_varp_data(A, Y0, U)

cf <- companion_format(Y, nu, A, U)

cf$U
cf$Y - cf$A \%*\% cf$Z

\dontrun{
# input has to be in matrix form
 companion_format(Y[1, ], nu[1, ], A[1, ], U[1,])
}

big_A(A)

Y <- matrix(seq_len(K * N), K, N)
big_Y(Y, p)

big_nu(nu, p)

\dontrun{

big_nu(c(nu), p)
}

U <- matrix(seq_len(K * N), K, N)
big_U(U, p)

SIGMA <- matrix(0.5, K, K)
big_SIGMA(SIGMA, p = p)
}
