% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-1-simulation.R
\name{create_arp_data}
\alias{create_arp_data}
\title{Create data using an AR(1)}
\usage{
create_arp_data(a, y0, e = rnorm(N), N = 100, intercept = FALSE)
}
\arguments{
\item{a}{A numeric vector. It specifies the lag coefficients and its length
will determine the maximum lag order \eqn{p}.}

\item{y0}{A numeric vector. It provides pre-sample observations and its
length must be at minimum \eqn{p} (see above).}

\item{e}{A numeric vector. It contains the innovations that hit the system.
If unspecified, it will draw from a standard normal distribution.}

\item{N}{An integer scalar. The sample size. Defaults to \code{N = 100}.}

\item{intercept}{A logical scalar. If the first entry of \code{a} is the value of
the intercept, set to \code{TRUE}. Otherwise leave at \code{FALSE}.}
}
\value{
An numeric vector of length \code{N}.
}
\description{
Given some starting values, coefficients and a sequence of error vectors,
\code{create_arp_data} will compute a sequence of observables through a
univariate autoregressive model.
}
