% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-1-simulation.R
\name{create_var1_data}
\alias{create_var1_data}
\title{Create data using a VAR(1)}
\usage{
create_var1_data(A, Y_0, EE)
}
\arguments{
\item{A}{A (K x K) matrix, providing the coeffcients for lag one.}

\item{Y_0}{A (K x 1) vector or column matrix, will be used as starting
values.}

\item{EE}{A (K x T) matrix, providing the sequence of error vectors.}
}
\value{
A (K x T) matrix of observables. The first column will equal \code{A
  \%*\% Y_0 + EE[, 1]}. The final observation of the K variables will be in
column T.
}
\description{
Given some starting values, coefficients and a sequence of error vectors,
\code{create_var1_data} will compute a sequence of observables according to a
vector autoregressive model of order one.
}
\section{Note}{
 For a faster implementation, see
\href{http://gallery.rcpp.org/articles/simulating-vector-autoregressive-process/}{this
solution} by Dirk Eddelbuettel.
}

