% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{dm_col_index}
\alias{dm_col_index}
\title{Column indices of duplication matrix}
\usage{
dm_col_index(K)
}
\arguments{
\item{K}{An integer scalar. The number of variables of the VAR(p) system.}
}
\value{
An integer vector of length K^2.
}
\description{
Create a vector containing the column indices of a duplication matrix. It's
assumed the output will be matched with a row index running from 1, 2, ...
onwards. It's helpful for creating a duplication matrix for a base matrix of
dimension K x K.
}
\details{
\code{dm_col_index} is a helper function for \code{\link{duplication_matrix}}.
}
