% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVAR-3-identification.R
\name{duplication_matrix_ginverse}
\alias{duplication_matrix_ginverse}
\title{Moore-Penrose inverse of duplication matrix \code{D}}
\usage{
duplication_matrix_ginverse(K)
}
\arguments{
\item{K}{Scalar integer, number of variables in the system.}
}
\value{
A \code{(K * (K + 1) / 2 x K^2)} matrix.
}
\description{
Simple wrapper of \code{\link[MASS:ginv]{MASS::ginv()}}.
}
