% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{expander_e}
\alias{expander_e}
\title{Expand a Matrix}
\usage{
expander_e(r, c = 1)
}
\arguments{
\item{r}{An integer scalar. It specifies the number of rows to create.}

\item{c}{An integer scalar. It specifies the number of columns to create.

\code{r} and \code{c} must be equal or \code{c} must be a one.}
}
\value{
An \code{(r x c)} matrix. The very first element in the
top left corner is a one, the rest are zeros.
}
\description{
Create a vector or matrix for expanding the dimension of another matrix by
adding rows and columns filled with zeros. Multiply the result of
\code{expander_e} with a matrix using the Kronecker product. See examples.
}
\examples{

expander_e(3)
expander_e(4, 4)

expander_e(3, 3) \%x\% matrix(0.1, 4, 4)

\dontrun{

expander_e(4, 3)
}
}
