% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{lag_length}
\alias{lag_length}
\title{Retrieve number of lags}
\usage{
lag_length(A)
}
\arguments{
\item{A}{A \code{(K x Kp)} matrix, providing the coeffcients for lag \code{1} to \code{p}
with the first row containing the coefficents of the first equation.
Parameter \code{p} is the maximum lag length and \code{K} the number of variables.}
}
\value{
An integer scalar. The lag length \code{p}.
}
\description{
Given a matrix of slope parameters, retrieve the number of lags.
}
\details{
At the moment, this function is just an intentionally named wrapper of
\code{ncol(A) / nrow(A)}.
}
\examples{
p <- 3
K <- 4
A <- matrix(0, K, K * p)
lag_length(A) == p

\dontrun{

A <- matrix(0, 4, 7)
lag_length(A)
lag_length(A[1, ])
}
}
