% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-2-estimation.R
\name{mle_var}
\alias{mle_var}
\title{Maximum likelihood estimation of a VAR(p)}
\usage{
mle_var(Y, p, init, log_lik = log_lik_init(Y, p), gradient)
}
\description{
Maximum likelihood estimation of a VAR(p)
}
\examples{
K <- 3
N <- 5E2
p <- 2

set.seed(8191)

A <- matrix(0.1, K, K * p)
Y0 <- matrix(0, K, p)
U <- matrix(rnorm(K * N), K, N)

Y <- create_varp_data(A, Y0, U)

mu <- rep(0, K)
A <- matrix(0.1, K, K * p)
SIGMA <- matrix(0, K, K)
diag(SIGMA) <- 1

args <- c(mu = mu, a = vec(A), s = vech(SIGMA))

ols_fit <- ols_mv(Y = Y, p = p, const = TRUE)

mle_fit <- mle_var(Y, p)

mle_fit$BETA.hat
ols_fit$BETA.hat

mle_fit$SIGMA.hat
ols_fit$SIGMA.hat

mle_fit$std.err
ols_fit$std.err
}
