% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVAR-2-estimation.R
\name{ols_cholesky}
\alias{ols_cholesky}
\alias{mle_svar}
\title{Estimate Contemporaneous Structural Effects}
\usage{
ols_cholesky(Y, p)

mle_svar(Y, p, By, Be)
}
\arguments{
\item{Y}{A \code{(K x N+p)} matrix carrying the data for estimation. There are
\code{N} observations for each of the \code{K} variables with \code{p} pre-sample values.}

\item{p}{An integer scalar. The lag length of the VAR(p) system.}

\item{By}{A \code{(K x K)} matrix. Determines the contemporaneous relation between
the endogeneous variables \code{y_t}. See examples and details below.}

\item{Be}{A \code{(K x K)} matrix. Determines the contemporaneous impact of a
structural shock \code{e_t} on the endogeneous variables \code{y_t}. See examples and
details below.}
}
\value{
The results may not necessarily be unique. The column sign, for
example, could be indeterminate.
\itemize{
\item \code{ols_cholesky} \cr A \code{(K x K)} numeric lower triangular matrix.
}

\itemize{
\item \code{mle_svar} \cr A list with two elements \code{By} and \code{Be}. Both are \code{(K x K)} numeric matrices. The elements which previously held \code{NA}s were
replaced with estimates.
}
}
\description{
Estimate structural matrices \code{By} and \code{Be} by least squares or
maximum likelihood.
}
\details{
The two matrices \code{By} and \code{Be} refer to the instantaneous effects matrices
in a vector autoregressive system \deqn{B_y y_t = A^*_1y_{t-1} + \dots +
A^*_py_{t-p} + B_e e_t.}{By * y_t = A_s(L) y_t + Be * e_t.} Thus, \code{By}
describes the contemporaneous effects between the elements of \code{y}. \code{Be}
describes the contemporaneous impact a structural error vector \code{e_t} has on
\code{y_t}.
\itemize{
\item \code{ols_cholesky} \cr Apply OLS and a Cholesky decomposition for recovering
\code{Be} from a simple recursive system. As usual, the ordering of the variables
will matter for a standard Cholesky decomposition. The effect matrix \code{By} is
assumed to be an identity matrix.
}

\itemize{
\item \code{mle_svar} \cr Estimate \code{By} and \code{Be} by maximising the
concentrated log-likelihood. The reduced-form parameters are first
estimated by OLS while the remaining structural parameters are estimated by
MLE in a second step. The function covers recursive and non-recursive,
exactly identified and over-identified systems. The necessary restrictions
on matrices \code{By} and \code{Be} are imposed by assigning numeric values to the
restricted elements. Unrestricted elements are left to be estimated by
setting them to \code{NA}.
}
}
\examples{
set.seed(8191)

# number of variables, observations and lag length
K <- 3
N <- 1E6
p <- 2

# prepare input
A <- cbind(matrix(0.1, K, K), matrix(-0.05, K, K)); diag(A) <- 0.4
Be <- matrix(0.4, K, K); Be[upper.tri(Be)] <- 0
Y0 <-matrix(0, nrow = K, ncol = p)
W <- matrix(rnorm(N * K), nrow = K, ncol = N)

# draw data and estimate Be
Y <- create_svar_data(A, Be, Y0, W)
Be_hat <- ols_cholesky(Y, p)
Be_hat
set.seed(8191)

K <- 3
N <- 1E6
p <- 2

A <- cbind(matrix(0.1, K, K), matrix(-0.05, K, K)); diag(A) <- 0.4
Be <- matrix(0.4, K, K); Be[upper.tri(Be)] <- 0
Y0 <-matrix(0, nrow = K, ncol = p)
W <- matrix(rnorm(N * K), nrow = K, ncol = N)

Y <- create_svar_data(A, Be, Y0, W)

By_init <- diag(K)
Be_init <- matrix(0, K, K)
Be_init[lower.tri(Be_init, diag = TRUE)] <- NA

mle_svar(Y, p, By_init, Be_init)
}
