% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-2-estimation.R
\name{ols_mv}
\alias{ols_mv}
\title{Multivariate ordinary least squares for VARs}
\usage{
ols_mv(Y, p, const = TRUE)
}
\arguments{
\item{Y}{A \code{(K x N)} matrix carrying the data for estimation. There are
\code{N} observations for each of the \code{K} variables.}

\item{p}{An integer scalar, the lag length used for estimation.}

\item{const}{A boolean scalar, indicating wether a constant should be
included. Defaults to \code{TRUE}.}
}
\value{
A list with four elements:
\itemize{
\item \code{BETA.hat} is a \code{(K x [K * p + 1])} or \code{(K x [K * p])} named matrix
containing the coefficient estimates. Its dimension depends on wether a
constant is included. Each column contains the coefficents for a particular
regressor, each row corresponds to a single equation. If \code{Y} did not
carry names, the variables will be named \emph{y1} and counting. The names of
the lagged regressors are derived by appending the variable name with a
\emph{.l1} for lag one etc.
If a constant is included, its coefficients are in the first column.
\item \code{SIGMA.hat} is a \code{(K x K)} named matrix. It is the covariance matrix of the
residuals. The columns and rows are named after the variables in \code{Y}.
\item \code{U.hat} is a \code{(K x N)} matrix of residuals. Its rows are named after the
variables, too.
\item \code{std.err} is a matrix of the same dimension and naming scheme as
\code{BETA.hat}. It has the standard errors of the coefficient estimates.
}
}
\description{
Perform multivariate OLS on a set of observables. Creation of lags for the
VAR will be handled by the function.
}
\details{
This routine applies equation-wise OLS to a VAR. It is equivalent to running
OLS with \code{p} lags for each of the variables and collecting the coefficients
and residuals in a matrix.
}
