% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVAR-3-identification.R
\name{rank_condition}
\alias{rank_condition}
\alias{rank_A_model}
\alias{rank_B_model}
\alias{rank_AB_model}
\title{Calculate matrix ranks for identification}
\usage{
rank_A_model(A, SIGMA_U = A_INV \%*\% t(A_INV),
  C_A = selection_matrix(A))

rank_B_model(B, C_B = selection_matrix(B))

rank_AB_model(A, B, SIGMA_U = A_INV \%*\% B \%*\% t(B) \%*\% t(A_INV),
  C_A = selection_matrix(A), C_B = selection_matrix(B))
}
\arguments{
\item{A}{A square numeric matrix, the coefficient of contemporaneous effects
between endogenous variables. Its dimension is (K x K).}

\item{SIGMA_U}{A square numeric matrix, the reduced-form residual
covariances. Its dimension is (K x K). The default setting assumes unit
variance of the structural shocks.}

\item{C_A, C_B}{A numeric matrix, the selection matrix for imposing linear
restrictions on matrix \code{A} or \code{B}. Its dimension is (K * (K + 1) / 2 x
K^2). The default assumes all elements of \code{A} or \code{B} equal to 0 or 1 are
restricted to those values.}

\item{B}{A square numeric matrix, the coefficient of contemporaneous effects
of structural shocks on endogenous variables. Its dimension is (K x
K).}
}
\value{
A scalar integer. The rank of the respective partial derivative
involved in achieving identification. For successful identification, the
rank must be equal to the number of unidentified parameters.
\itemize{
\item For the the A-type model, the rank has to be
\ifelse{latex}{\out{$K^2 + \frac{K}{2} (K + 1)$}}{K^2 + K * (K + 1) / 2}.
\item For the the B-type model, the rank has to be
\ifelse{latex}{\out{$K^2$}}{K^2}.
\item For the the AB-type model, the rank has to be
\ifelse{latex}{\out{$2K^2$}}{2K^2}.
}
}
\description{
There are necessary and sufficient algebraic conditions for checking whether
a model is (locally) identified. These conditions relate to the rank of
certain matrices of first-order partial derivatives. The following functions
compute those ranks.
}
\details{
Note that \code{rank_AB_model} nests both \code{rank_A_model} and \code{rank_B_model} if
supplied with the appropriate arguments.
}
\examples{
set.seed(819)
K <- 4
A <- matrix(rnorm(K^2), K, K)
A[sample((1:K^2)[-seq(1, K^2, K+1)], K * (K-1) / 2)] <- 0
diag(A) <- 1
A
SIGMA_U <- solve(A) \%*\% t(solve(A))
rank_A_model(A)
is_identified(A, SIGMA_U = SIGMA_U)

K <- 3
A <- matrix(rnorm(K^2), K, K)
A[c(3, 6, 7)] <- 0
A[c(1, 5, 9)] <- 1

}
