% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{selector_J}
\alias{selector_J}
\title{Create Selection Matrix \code{J}}
\usage{
selector_J(K, p)
}
\arguments{
\item{K}{An integer scalar. The number of variables of the VAR(p) system.}

\item{p}{An integer scalar. The lag length of the VAR(p) system.}
}
\value{
A \code{(K x Kp)} numerical matrix containing ones and zeros in selected
places. The first \code{(K x K)} elements form an identity matrix. The rest are
zeros.
}
\description{
Create the selection matrix \code{J} for converting objects in VAR(1) companion
format to their original VAR(p) representation.
}
\examples{
K <- 4
p <- 2

J <- zeitreihe:::selector_J(K, p)
}
