% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{var_length}
\alias{var_length}
\title{Retrieve number of variables}
\usage{
var_length(mat)
}
\arguments{
\item{mat}{\code{(K x M)} matrix, with \code{K} being the number of variables.}
}
\value{
An integer scalar. The number of variables \code{K}.
}
\description{
Given a matrix of observations, residuals or slope coefficients, retrieve the
number of variables \code{K}.
}
\details{
At the moment, this function is just an aptly / intentionally named wrapper of
\code{nrow(mat)}.
}
\examples{
p <- 3
K <- 4
var_length(matrix(0, K, K * p)) == K
}
