% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{vec}
\alias{vec}
\title{Vectorise a matrix}
\usage{
vec(mat)
}
\arguments{
\item{mat}{An \code{(M x N)} matrix with arbitrary dimensions \code{M} and \code{N}.}
}
\value{
A column matrix of dimension \code{NM x 1}.
}
\description{
Vectorise a matrix by stacking its columns in a vector.
}
\examples{
mat <- matrix(1:15, 3, 5)
vec(mat)
all.equal(mat, matrix(vec(man), 3, 5))
}
