% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{vech}
\alias{vech}
\title{Vectorise a symmetric matrix}
\usage{
vech(mat)
}
\arguments{
\item{mat}{An \code{(M x M)} matrix with arbitrary dimensions \code{M}.}
}
\value{
A column matrix of dimension \code{(M^2 + M)/2 x 1}.
}
\description{
Vectorise a symmetric matrix by stacking the elements on its lower
triangular matrix including the diagonal.
}
\examples{
mat <- matrix(1:8, 4, 4)
mat[upper.tri(mat)] <- t(mat)[upper.tri(mat)]
vech(mat)
}
