% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{Y2Z}
\alias{Y2Z}
\title{Transform a matrix of time-series by adding lags}
\usage{
Y2Z(Y, p)
}
\arguments{
\item{Y}{A (K x T) matrix, containing the data of the K variables.}

\item{p}{An integer, the number of lags to create. Must be greater or equal 0.}
}
\value{
A ((K * (p + 1)) x (T - p)) matrix, the first K rows will contain the
 original data and the remaining rows will contain its lags up to order p.
 The columns will contain time periods, of which there are now T - p left
 after setting aside pre-sample values for the lags.
}
\description{
A matrix Y is transformed into Z notation (see Lütkepohl (2005, p. 70)). This
facilitates estimation of VAR(p) models.
}
\section{Note}{
 For regression purposes, the first K rows would have to be
 discarded as they contain Y itself; see example.
}

\section{ToDo}{

add sanity checks... data.frame? matrix? dimension? max, min p?
no intercept (yet)!
rownames?
}

\examples{
Y <- draw_data_DH08_m1(n = 5e5)

X <- Y2Z(Y, p = 4)
Y <- X[seq_len(4), ] # has correct sample size
Z <- X[-seq_len(4), ]

B <- Y \%*\% t(Z) \%*\% solve(Z \%*\% t(Z))

}
