% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-1-simulation.R
\name{create_varp_data}
\alias{create_varp_data}
\title{Create data using a VAR(p)}
\usage{
create_varp_data(B, Z_0, UU)
}
\arguments{
\item{B}{A (K x Kp) matrix, providing the coeffcients for lag 1 - p.}

\item{Z_0}{A (K x p)  matrix, will be used as starting values.}

\item{UU}{A (K x T) matrix, providing the sequence of error vectors.}
}
\value{
A (K x T) matrix of observables. The first columns will be equal to
\code{Z_0}. The final observation of the K variables will be in column T.
}
\description{
Given some starting values, coefficients and a sequence of error vectors,
\code{create_varp_data} will compute a sequence of observables according to a
vector autoregressive model of order p.
}
\section{Note}{
 For a faster implementation, see
\href{http://gallery.rcpp.org/articles/simulating-vector-autoregressive-process/}{this
solution} by Dirk Eddelbuettel.
}

