% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-1-simulation.R
\name{create_varp_data}
\alias{create_varp_data}
\title{Create data using a reduced-form VAR}
\usage{
create_varp_data(A, Y0, U)
}
\arguments{
\item{A}{A \code{(K x Kp)} matrix, providing the coeffcients for lag \code{1} to \code{p}
with the first row containing the coefficents of the first equation.
Parameter \code{p} is the maximum lag length and \code{K} the number of variables.}

\item{Y0}{A \code{(K x p)} matrix which will be used as starting values. The
first column corresponds to the very first time period.}

\item{U}{A \code{(K x N)} matrix, providing the sequence of error vectors.}
}
\value{
A \code{(K x (N + p))} matrix holding the observations. The first \code{p}
columns will be equal to \code{Y0}. Column \code{p + 1} will be equal to \code{A \\\\\%*\\\\\% Y0 + U[, 1]}, where \code{U} contains reduced form errors. The final observation
of the \code{K} variables will be in column \code{N}.
}
\description{
Given some starting values, coefficients and a sequence of error vectors,
\code{create_varp_data} will compute a sequence of observables using a simple
vector autoregressive process.
}
\section{Note}{
 For a faster implementation, see
\href{http://gallery.rcpp.org/articles/simulating-vector-autoregressive-process/}{this solution}
by Dirk Eddelbuettel.
}

\seealso{
Other functions for creating data: \code{\link{create_svar_data}}
}
\concept{functions for creating data}
