% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{mean_var_process}
\alias{mean_var_process}
\title{Compute the mean of a VAR(p) process}
\usage{
mean_var_process(A, nu)
}
\arguments{
\item{A}{A \code{(K x Kp)} matrix, providing the coeffcients for lag \code{1} to \code{p}
with the first row containing the coefficents of the first equation.
Parameter \code{p} is the maximum lag length and \code{K} the number of variables.}

\item{nu}{A \code{(K x 1)} matrix of intercepts.}
}
\value{
A \code{(K x 1)} matrix containing the unconditional mean of each
variable.
}
\description{
Compute the mean of a VAR(p) process
}
\examples{
 A <- matrix(c(.5, .4, .1, .5, 0, .25, 0, 0), nrow = 2)
 nu <- c(2, 3)
 mean_var_process(A, nu)

\dontrun{

 A <- matrix(c(1, 0, 0, 1), nrow = 2)
 nu <- c(2, 3)
 mean_var_process(A, nu)
 }
}
