% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-2-estimation.R
\name{ols_mv}
\alias{ols_mv}
\title{Multivariate ordinary least squares}
\usage{
ols_mv(data, p)
}
\arguments{
\item{data}{A matrix, ..}

\item{p}{An integer, the largest lag.}
}
\value{
A list with four components \code{BETA.hat}, \code{SIGMA.hat},
\code{U.hat}, and \code{std.err}. The first three are vectors, the last
component is a scalar.

\code{BETA.hat} contains the coefficient estimates, ...
}
\description{
Multivariate ordinary least squares (<=> equation-wise ols)
}
