% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVAR-3-identification.R
\name{rank_condition}
\alias{rank_condition}
\alias{rank_A_model}
\alias{rank_B_model}
\title{Calculate ranks for identification}
\usage{
rank_A_model(A, C_A = selection_matrix(A), SIGMA_U = A_INV \%*\%
  t(A_INV))

rank_B_model(B, C_B = selection_matrix(B))
}
\arguments{
\item{A}{A square numeric matrix, the coefficient of contemporaneous effects
between endogenous variables. Its dimension is (K x K).}

\item{C_A, C_B}{A numeric matrix, the selection matrix for imposing linear
restrictions on matrix \code{A} or \code{B}. Its dimension is (K * (K + 1) / 2 x
K^2). The default assumes all elements of \code{A} or \code{B} equal to 0 or 1 are
restricted to those values.}

\item{SIGMA_U}{A square numeric matrix, the reduced-form residual
covariances. Its dimension is (K x K). The default setting assumes unit
variance of the structural shocks.}

\item{B}{A square numeric matrix, the coefficient of contemporaneous effects
of structural shocks on endogenous variables. Its dimension is (K x
K).}
}
\value{
A scalar integer. The rank of the respective partial derivative
involved in achieving identification.
\itemize{
\item For the the A-type model, the rank has to be
\ifelse{latex}{\out{$K^2 + \frac{K}{2} (K + 1)$}}{K^2 + K * (K + 1) / 2}.
\item For the the B-type model, the rank has to be
\ifelse{latex}{\out{$K^2$}}{K^2}.
}
}
\description{
There are necessary and sufficient algebraic conditions for checking whether
a model is (locally) identified. These conditions relate to the rank of
certain matrices of first-order partial derivatives. The following functions
compute those ranks.
}
