% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVAR-3-identification.R
\name{selection_matrix}
\alias{selection_matrix}
\title{Selection matrix for imposing linear restrictions}
\usage{
selection_matrix(X, value = 0:1, flatten = vec)
}
\arguments{
\item{X}{A square matrix, the target of imposing restrictions. The restricted
elements must coincide with the values specified for \code{value} below. The
non-restricted elements may have \code{NA}s or any other value except those in
\code{value}.}

\item{value}{A numeric vector, the values imposed on elements of \code{X}. May be
numeric or contain \code{NA}. Every element of \code{X} with any of these values will
be assumed to be restricted to that value.}

\item{flatten}{A function, flattening the dimension of \code{X}. The default is to
\code{\link[=vec]{vec()}}, but in some situations may want to select unique elements of a
symmetric matrix and then \code{\link[=vech]{vech()}} is helpful.}
}
\value{
A matrix containing 0s and 1s of dimension
\itemize{
\item R x K^2
}

if \code{flatten = vec} or
\itemize{
\item R x (K^2 + K)/2
}

if \code{flatten = vech}. The scalar R is the number of restrictions and K the
number of variables, i.e. K is the number of both the rows and the columns of
\code{X}.
}
\description{
Selection matrix selecting those elements in \code{X} with restrictions in place.
This function will create the matrix \code{C} in \code{C \%*\% vec(X) = c}.
}
\note{
This function is only able to create selection matrices for
restrictions on individual elements. Restrictions on linear combinations of
elements are not (yet) implemented.
}
\examples{
set.seed(8191)
K <- 4
A <- matrix(rnorm(K^2), K, K)
A[sample(1:K^2, K * (K-1) / 2)] <- 0
selection_matrix(A)

}
