% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR-2-estimation.R
\name{sMA_CI}
\alias{sMA_CI}
\title{Hall's percentile interval bootstrap}
\usage{
sMA_CI(Y, p, h, alpha, reps)
}
\arguments{
\item{reps}{}
}
\description{
Hall's percentile interval bootstrap
}
\examples{
data("Canada", package = "vars")
B <- zeitreihe::ols_cholesky(Y = t(Canada), p = 2)
IRFs <- zeitreihe::sMA_CI(Y = t(Canada), 2, 20, 0.1, 100)

}
