% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_funs.R
\name{Y2Z}
\alias{Y2Z}
\title{Create a regressor matrix for a VAR(p) model}
\usage{
Y2Z(Y, p, const = TRUE)
}
\arguments{
\item{Y}{A \code{(K x N)} matrix carrying the data for estimation. There are
\code{N} observations for each of the \code{K} variables.}

\item{p}{An integer scalar, the lag length used for estimation.}

\item{const}{A boolean scalar, indicating wether a constant should be
included. Defaults to \code{TRUE}.}
}
\value{
A \code{([K * p + 1] x [N - p])} or \code{([K * p] x [N - p])}matrix. The
dimension depends on whether a constant is included. If it is included, the
first row consists of ones. The next K rows will contain the data lagged by
one time period. The remaining rows will contain further lags up to order p.
The columns will contain time periods of which there are now \code{N - p} left
after setting aside pre-sample values for the lags.
}
\description{
Matrix \code{Y} is transformed into \code{Z} notation; see Lütkepohl (2005, p. 70).
This transformation facilitates estimation of VAR(p) models.
}
\examples{
K <- 3
N <- 6
p <- 2

Y <- matrix(seq_len(K*N), nrow = K)
Z <- Y2Z(Y, p)
}
